+++ feedme/__main__.py
@@ -23,4 +23,4 @@
 from .cli import cli
 
 if __name__ == '__main__':
-    cli(prog_name='python -m aggregator')
+    cli(prog_name='python -m feedme')